fixed redux and storage interaction
the initial application of storage use and redux was not working because i forgot to add items to the storage initially. now each time a new category is added items are taken out of storage, a new category is added to items and they are sent back to storage again. this way the storage and the client are always in sync