save notes on popup close, refactoring
i'm using visibilityChange event that is placed in a hook to track when the popup is being closed. Chrome doesn't provide any tools for tracking the popup closing. I also refactored some of the code and i'm only doing checks in react if the editors are open or closed. None of the data goes to the storage. I also switched to using chrome's local storage as sync storage was running out of space