visibility change, refactoring
i was having trouble figuring out how to listen for the chrome popup closing as chrome doesn't provide any methods for that. one solution is to listen for a visibility change event because it is the last event that is fired if the tab is closed or put in background. i've added it to an effect hook where i'm listening for the event and dispatching the required actions